refactor: co-locate each component with its test in its own folder
Move SkipLink and BubbleNav into per-component folders, each holding the
component file and its test (no index file; imports point at the component
file directly). icons.tsx stays a shared module.
