+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter, JetBrains_Mono } from "next/font/google";
-import { SkipLink } from "@/components/SkipLink";
-import { BubbleNav } from "@/components/nav/BubbleNav";
+import { SkipLink } from "@/components/SkipLink/SkipLink";
+import { BubbleNav } from "@/components/nav/BubbleNav/BubbleNav";
 import { site } from "@/lib/content";
 import "./globals.css";
